Moved doMove and undoMove to move.cpp and move.h. Reformatted all code files. Fixed misinterpretation of 6-bits origin and target values in move struct from x- and y-coordinates to shifts. Fixed wrong variable names in iteration in empty_between function. Set xDiff and yDiff to 1 in empty_between function, when the corresponding values of xChanged and yChanged are true to avoid immediate termination of the following for-loop. Fixed validity check for double-forward move of pawns that previously caused only white pieces to be allowed to perform double-forward moves. Fixed faulty resetting of the boards in undoMove function. Added printMove function to print a simple representation of a t_move object. Fixed type mismatch from int to uint64_t in board_value_from_shift function in util.cpp causing shifted values to overflow. Added length check for List array when freeing to avoid freeing unallocated memory.

+++ util.cpp
@@ -1,6 +1,5 @@
 #include "util.h"
 #include <stdexcept>
-#include <cstring>
 
 
 /* #############################################################################
@@ -164,6 +163,7 @@ Position Offset::toPosition() const {
 
 int shift_from_position(Position position) {
     return 63 - (position.y * 8 + (7 - position.x));
+    // return 63 - (position.y * 8 + position.x);
 }
 
 
@@ -172,13 +172,14 @@ Position position_from_shift(int shift) {
 
     int y = val / 8;
     int x = 7 - (val - (y * 8));
+    // int x = val - (y * 8);
 
     return Position {x, y};
 }
 
 
 bool board_value_from_shift(uint64_t board, int shift) {
-    if ((board & 1 << shift ) != 0)
+    if ((board & ((uint64_t) 1 << shift)) != 0)
         return true;
     return false;
 }
